requirements.txt
amqp==5.0.6
asgiref==3.2.10
beautifulsoup4==4.9.3
billiard==3.6.4.0
celery==5.1.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
Django==3.1
django-cors-headers==3.7.0
django-filter==2.4.0
django-phonenumber-field==5.0.0
djangorestframework==3.12.4
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
kombu==5.1.0
lxml==4.6.3
mailjet-rest==1.3.4
MarkupSafe==1.1.1
packaging==20.9
phonenumbers==8.12.19
Pillow==8.2.0
prompt-toolkit==3.0.18
psycopg2-binary==2.8.5
pyparsing==2.4.7
pytz==2021.1
redis==3.5.3
requests==2.25.1
ruamel.yaml==0.17.2
ruamel.yaml.clib==0.2.2
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
transliterate==1.10.2
uritemplate==3.0.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
